+++ src/Endpoints/MailAliases.php
@@ -3,13 +3,12 @@
 namespace Vdhicts\Cyberfusion\ClusterApi\Endpoints;
 
 use Vdhicts\Cyberfusion\ClusterApi\Exceptions\RequestException;
-use Vdhicts\Cyberfusion\ClusterApi\Models\MailAccount;
-use Vdhicts\Cyberfusion\ClusterApi\Models\MailAccountUsage;
+use Vdhicts\Cyberfusion\ClusterApi\Models\MailAlias;
 use Vdhicts\Cyberfusion\ClusterApi\Request;
 use Vdhicts\Cyberfusion\ClusterApi\Response;
 use Vdhicts\Cyberfusion\ClusterApi\Support\ListFilter;
 
-class MailAccounts extends Endpoint
+class MailAliases extends Endpoint
 {
     /**
      * @param ListFilter|null $filter
@@ -24,7 +23,7 @@ class MailAccounts extends Endpoint
 
         $request = (new Request())
             ->setMethod(Request::METHOD_GET)
-            ->setUrl(sprintf('mail-accounts?%s', http_build_query($filter->toArray())));
+            ->setUrl(sprintf('mail-aliases?%s', http_build_query($filter->toArray())));
 
         $response = $this
             ->client
@@ -34,9 +33,9 @@ class MailAccounts extends Endpoint
         }
 
         return $response->setData([
-            'mailAccounts' => array_map(
+            'mailAliases' => array_map(
                 function (array $data) {
-                    return (new MailAccount())->fromArray($data);
+                    return (new MailAlias())->fromArray($data);
                 },
                 $response->getData()
             )
@@ -52,7 +51,7 @@ class MailAccounts extends Endpoint
     {
         $request = (new Request())
             ->setMethod(Request::METHOD_GET)
-            ->setUrl(sprintf('mail-accounts/%d', $id));
+            ->setUrl(sprintf('mail-aliases/%d', $id));
 
         $response = $this
             ->client
@@ -62,54 +61,30 @@ class MailAccounts extends Endpoint
         }
 
         return $response->setData([
-            'mailAccount' => (new MailAccount())->fromArray($response->getData()),
+            'mailAlias' => (new MailAlias())->fromArray($response->getData()),
         ]);
     }
 
     /**
-     * @param int $id
+     * @param MailAlias $mailAlias
      * @return Response
      * @throws RequestException
      */
-    public function usages(int $id): Response
-    {
-        $request = (new Request())
-            ->setMethod(Request::METHOD_GET)
-            ->setUrl(sprintf('mail-accounts/usages/%d', $id));
-
-        $response = $this
-            ->client
-            ->request($request);
-        if (! $response->isSuccess()) {
-            return $response;
-        }
-
-        return $response->setData([
-            'mailAccountUsage' => (new MailAccountUsage())->fromArray($response->getData()),
-        ]);
-    }
-
-    /**
-     * @param MailAccount $mailAccount
-     * @return Response
-     * @throws RequestException
-     */
-    public function create(MailAccount $mailAccount): Response
+    public function create(MailAlias $mailAlias): Response
     {
         $requiredAttributes = [
             'localPart',
-            'password',
+            'forwardEmailAddresses',
             'mailDomainId',
         ];
-        $this->validateRequired($mailAccount, 'create', $requiredAttributes);
+        $this->validateRequired($mailAlias, 'create', $requiredAttributes);
 
         $request = (new Request())
             ->setMethod(Request::METHOD_POST)
-            ->setUrl('mail-accounts')
-            ->setBody($this->filterFields($mailAccount->toArray(), [
+            ->setUrl('mail-aliases')
+            ->setBody($this->filterFields($mailAlias->toArray(), [
                 'local_part',
-                'password',
-                'quota',
+                'forward_email_addresses',
                 'mail_domain_id',
             ]));
 
@@ -121,33 +96,32 @@ class MailAccounts extends Endpoint
         }
 
         return $response->setData([
-            'mailAccount' => (new MailAccount())->fromArray($response->getData()),
+            'mailAlias' => (new MailAlias())->fromArray($response->getData()),
         ]);
     }
 
     /**
-     * @param MailAccount $mailAccount
+     * @param MailAlias $mailAlias
      * @return Response
      * @throws RequestException
      */
-    public function update(MailAccount $mailAccount): Response
+    public function update(MailAlias $mailAlias): Response
     {
         $requiredAttributes = [
             'localPart',
-            'password',
+            'forwardEmailAddresses',
             'mailDomainId',
             'id',
             'clusterId'
         ];
-        $this->validateRequired($mailAccount, 'update', $requiredAttributes);
+        $this->validateRequired($mailAlias, 'update', $requiredAttributes);
 
         $request = (new Request())
             ->setMethod(Request::METHOD_PUT)
-            ->setUrl(sprintf('mail-accounts/%d', $mailAccount->id))
-            ->setBody($this->filterFields($mailAccount->toArray(), [
+            ->setUrl(sprintf('mail-aliases/%d', $mailAlias->id))
+            ->setBody($this->filterFields($mailAlias->toArray(), [
                 'local_part',
-                'password',
-                'quota',
+                'forward_email_addresses',
                 'mail_domain_id',
                 'id',
                 'cluster_id',
@@ -161,7 +135,7 @@ class MailAccounts extends Endpoint
         }
 
         return $response->setData([
-            'mailAccount' => (new MailAccount())->fromArray($response->getData()),
+            'mailAlias' => (new MailAlias())->fromArray($response->getData()),
         ]);
     }
 
@@ -174,7 +148,7 @@ class MailAccounts extends Endpoint
     {
         $request = (new Request())
             ->setMethod(Request::METHOD_DELETE)
-            ->setUrl(sprintf('mail-accounts/%d', $id));
+            ->setUrl(sprintf('mail-aliases/%d', $id));
 
         return $this
             ->client