+++ src/Models/MailDomain.php
@@ -8,6 +8,7 @@ use Vdhicts\Cyberfusion\ClusterApi\Contracts\Model;
 class MailDomain implements Model
 {
     public string $domain;
+    public array $catchAllForwardEmailAddresses = [];
     public int $unixUserId;
     public ?int $id = null;
     public ?int $clusterId = null;
@@ -18,6 +19,7 @@ class MailDomain implements Model
     {
         $mailDomain = new self();
         $mailDomain->domain = Arr::get($data, 'domain');
+        $mailDomain->catchAllForwardEmailAddresses = Arr::get($data, 'catch_all_forward_email_addresses', []);
         $mailDomain->unixUserId = Arr::get($data, 'unix_user_id');
         $mailDomain->id = Arr::get($data, 'id');
         $mailDomain->clusterId = Arr::get($data, 'cluster_id');
@@ -30,6 +32,7 @@ class MailDomain implements Model
     {
         return [
             'domain' => $this->domain,
+            'catch_all_forward_email_addresses' => $this->catchAllForwardEmailAddresses,
             'unix_user_id' => $this->unixUserId,
             'id' => $this->id,
             'cluster_id' => $this->clusterId,