+++ CHANGELOG.md
@@ -2,7 +2,27 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to 
-[Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+[Semantic Versioning](https://semver.org/spec/v2.0.0.html). Please note this changelog affects this package and not the 
+cluster API. See the changelog of the [cluster API](https://cluster-api.cyberfusion.nl/redoc#section/Changelog) for 
+detailed information.
+
+## [1.6.0]
+
+### Added
+
+- Add database users endpoint.
+- Add mail aliases endpoint.
+- Add `balancer_backend_name` to virtualhosts.
+- Add `catch_all_forward_email_addresses` to mail domain.
+- Add commit call to the cluster endpoint. See the [readme](readme.md) for more information.
+
+### Changed
+
+- Update to [API version 1.12](https://cluster-api.cyberfusion.nl/redoc#section/Changelog/1.12-2021-02-23)
+
+### Removed
+
+- Remove `forward_email_addresses` from mail accounts.
 
 ## [1.5.0]